+++ addons/mail/static/src/views/activity/activity_record.js
@@ -5,6 +5,7 @@ import { Component } from "@odoo/owl";
 import { useService } from "@web/core/utils/hooks";
 import { Field } from "@web/views/fields/field";
 import {
+    getFormattedRecord,
     getImageSrcFromRecordInfo,
     isHtmlEmpty,
 } from "@web/views/kanban/kanban_record";
@@ -29,18 +30,13 @@ export class ActivityRecord extends Component {
             isHtmlEmpty,
         };
         const { arch, templateDocs } = this.props.archInfo;
-        const compileParams = {
-            recordExpr: `record`,
-        };
-        this.recordTemplate = useViewCompiler(ActivityCompiler, arch, templateDocs, compileParams)[
-            "activity-box"
-        ];
+        this.recordTemplate = useViewCompiler(ActivityCompiler, arch, templateDocs)["activity-box"];
     }
 
     getRenderingContext() {
         const { record } = this.props;
         return {
-            record: record.formattedRecord,
+            record: getFormattedRecord(record),
             activity_image: (...args) => getImageSrcFromRecordInfo(record, ...args),
             user_context: this.user.context,
             widget: this.widget,
